Fix direct access

+++ module/config/language.php
@@ -1,8 +1,9 @@
 <?php
 
+	if(!defined('COREINC')) die('Direct access not allowed');
 	$i18n = $app->apiLoad('coreI18n')->languageSet('fr')->load('config');
 
-// Remove
+	// Remove
 	if(sizeof($_POST['del']) > 0){
 		foreach($_POST['del'] as $e){
 			$app->dbQuery("DELETE FROM k_country WHERE iso='".$e."'");
